+++ lib/modules/requests/request_filter.ts
@@ -1,4 +1,4 @@
-import { Request } from 'puppeteer';
+import { Request, ResourceType } from 'puppeteer';
 import { matchText } from '../../utils/utils';
 import { ExpectedHeaders } from './types';
 
@@ -100,6 +100,14 @@ export default class RequestFilter {
         const requests = filterPromise(this.requests, el => {
             const response = el.response();
             return !response;
+        });
+        return new RequestFilter(requests);
+    }
+
+    public resourceType(type: ResourceType): RequestFilter {
+        const requests = filterPromise(this.requests, el => {
+            const resourceType = el.resourceType();
+            return resourceType === type;
         });
         return new RequestFilter(requests);
     }